move git things to a new file
